+++ package.json
@@ -53,11 +53,13 @@
   },
   "devDependencies": {
     "@testing-library/react": "^14.0.0",
+    "@types/redux-mock-store": "^1.0.3",
     "@types/uuid": "^9.0.1",
     "cypress": "^12.17.1",
     "eslint-plugin-cypress": "^2.13.3",
     "fetch-mock": "^9.11.0",
     "jest": "^27.5.1",
+    "jest-fetch-mock": "^3.0.3",
     "react-test-renderer": "^18.2.0",
     "redux-devtools-extension": "^2.13.9",
     "redux-mock-store": "^1.5.4"
